Rename SIMPLE to PIMPLE and add an additional inner loop. (#29150)
